agents/knowledge-agent/requirements.txt
## Python ML Dependencies (for knowledge-agent-env)
azure-identity==1.25.1
azure-search-documents==11.6.0
openai==2.8.1
fastapi[all]==0.122.0
numpy==2.3.5
pytest==9.0.1
redis==7.1.0
sentence-transformers==5.1.2
uvicorn[standard]==0.38.0